renamed classes and paths
